Fixed 4 features
