+++ Projects/RMAutomationUIFramework/src/main/java/org/fundacionjala/automation/framework/pages/admin/resource/ResourcePage.java
@@ -58,6 +58,8 @@ public class ResourcePage extends AdminPage {
 	}
 	
 	public boolean verifyResourceExist(String resourceName){
+
+		
 		if (verifyExist(resourceName)) {
 				LogManager.info("[TRUE] Resource " + resourceName + " exists");
 				return true;
@@ -220,7 +222,7 @@ public class ResourcePage extends AdminPage {
 
 	public ResourcePage clickOnLastPageButton() {
 		UIActions.waitFor(ResourceMap.LAST_PAGE_BUTTON);
-		UIActions.clickAt(firstPageButton);
+		UIActions.clickAt(lastPageButton);
 		return new ResourcePage();
 	}
 
@@ -229,6 +231,7 @@ public class ResourcePage extends AdminPage {
 		String lastPage = inputNumberPage.getAttribute("value").trim();
 		String xpath = ResourceMap.TOTAL_NUMBER_PAGE.replace("totalPages", lastPage);
 		try {
+			//UIActions.waitFor(xpath);
 			if(BrowserManager.getDriver().findElement(By.xpath(xpath))!= null)
 				return true;
 			else
