+++ Projects/RMAutomationUIFramework/src/test/resources/features/admin/resources.feature
@@ -1,46 +1,35 @@
 Feature: Resources
 	I want to create, update, remove and search resources
 
+#scenario 1
+Scenario: Create a new resource
+Given I as Administrator Login to Room Manager
+ When I Create a new resource with name "computer", display name "Computer", description "this is a new computer" and icon "fa-desktop"
+ Then Validate that the resource with name "computer" is diplayed in resource page
 
-Scenario Outline: Create a new resource
-Given I Login to Room Manager
-When I Create a new resource with name "<name>", with display name "<displayName>", with description "<description>" and icon "<icon>"
-Then I validate that the resource with name "<name>" is diplayed in resource page
-
-Examples:
-	|name		 |displayName	|description					 |icon			|
-	|computer|Computer	  |this is a new computer|fa-desktop|
-	
-#scenario 2
-Scenario Outline: Delete Resource
-Given I have a resource created with the name "<name>"
-When I delete the resource with the name "<name>"
-And Confirm the changes
-Then I verify that the resource with the name "<name>" has been deleted
-
-Examples:
-    | name  |
-    | TestResource  |
-
-#scenario 3
+ #scenario 2
+Scenario: Delete Resource
+Given I have a resource created with the name "testResource", display name "TestResource", description "TestResource" and icon "fa-fire"
+  And I as Administrator Login to Room Manager 
+ When I delete the resource with the name "testResource"
+  And Confirm the changes
+ Then Validate that the resource with the name "testResource" has been deleted
+ 
+ #scenario 3
 Scenario: A resource is displayed when it is filtered.
-Given I have a resource
-When I search the resource in resources page
-Then I validate the resource is displayed 
-	And then the resource is deleted
-
+Given I have a resource created with the name "Resource12345", display name "Resource12345", description "Resource12345>" and icon "fa-fire"
+  And I as Administrator Login to Room Manager
+ When I search the resource "Resource12345" in resources page
+ Then Validate the resource "Resource12345" is displayed 
+ 
 #scenario 4	
-Scenario Outline: Update Resource
-Given I have a resource created with name "<name>"
-When I modify the "<field>" field with value "<value>"
+Scenario: Update Resource
+Given I have a resource created with the name "Fire", display name "Fire", description "Fire resource" and icon "fa-fire"
+  And I as Administrator Login to Room Manager
+ When I modify the "displayname" field with value "newFirevalue" in the resource "Fire"
 	And I save the modifications
-Then the resource is modified according the changes ("<field>" field with value "<value>")
-
-#Fields to be modified : name, displayname, icon
-Examples:
-    | name                | field  |value 														|
-    | Resource12345       | name   |ThisisalongstringThisisalongstring|
-    
+ Then Validate that the resource "Fire" is modified according the changes ("displayname" field with value "newFirevalue")
+ 
 #scenario 5
 Scenario Outline: All resources are displayed in resource table
 Given I Login to Room Manager application
@@ -58,15 +47,14 @@ Given I log in to Room Manager app
  When I create a resource with name "<name>", display name "<displayName>", description "<description>" and icon "<icon>"
  Then I validate that resource with "<name>", "<displayName>" and "<icon>" is displayed
 
-# Examples:
+ Examples:
 	|name		 |displayName	|description					 |icon			|
   |folder	 |folder 	    |this is a new resource|fa-folder|
-
- #scenario 7
- Scenario Outline: The number of total items is displayed in the resource table
- Given I Login to RoomManager APP
-  When I create "<number>" resources
-  Then I validate that total resources created are displayed in resource table 
+#scenario 7
+Scenario Outline: The number of total items is displayed in the resource table
+Given I Login to RoomManager APP
+ When I create "<number>" resources
+ Then I validate that total resources created are displayed in resource table 
  Examples:
   	|number|
 		|		1	 | 
@@ -78,18 +66,16 @@ Given I have a "<name>" resource created
  When I associate the resource "<name>" to a room "<room>"
   And I want to remove the resource "<name>"
  Then I validate that the association is displayed
-
  Examples:
 	 	|name		 	 |room	|
-		|dashboard |Room01|
-	
+		|dashboard |Room001|
+
 #scenario 9
 Scenario Outline: The quantity selected in page size is displayed in resource table
 Given I have atleast "<numberResources>" resources created
-	And I Login to Room Manager
+  And I as Administrator Login to Room Manager
  When I select a option "<number>" on page size option
  Then I validate that the resource table size is same than the option "<number>" selected
-
  Examples:
  		|numberResources|number|
  		|				52	 		|	50	|
@@ -97,15 +83,15 @@ Given I have atleast "<numberResources>" resources created
 #scenario 10
 Scenario Outline: When First button is clicked the first page is displayed in resource table
 Given I have atleast "<numberResources>" created resources
-  And I Login to Room Manager
-When I Clicked on First button on resource table
+  And I as Administrator Login to Room Manager
+ When I Clicked on First button on resource table
  Then I validate that the first page is displayed on resource table
-
 Examples:  
 		|numberResources|
  		|				53	 	|	
  		
-#scenario 11
+ 		
+#scenario 11 
 Scenario Outline: When 'Last' button is clicked the last page is displayed in resource table
 Given There are atleast "<numberResources>" created resources
  When I Clicked on Last button on resource table
@@ -113,3 +99,5 @@ Given There are atleast "<numberResources>" created resources
  Examples:  
 		|numberResources|
  		|				15	 	|
+
+#scenario 12
